+++ src/iobuf.c
@@ -16,7 +16,7 @@
 #include <assert.h>
 #include "zenoh/iobuf.h"
 
-z_iobuf_t z_iobuf_wrap_wo(uint8_t *buf, unsigned int capacity, unsigned int rpos, unsigned int wpos)
+z_iobuf_t z_iobuf_wrap_wo(uint8_t *buf, size_t capacity, size_t rpos, size_t wpos)
 {
     assert(rpos <= capacity && wpos <= capacity);
     z_iobuf_t iobuf;
@@ -27,12 +27,12 @@ z_iobuf_t z_iobuf_wrap_wo(uint8_t *buf, unsigned int capacity, unsigned int rpos
     return iobuf;
 }
 
-z_iobuf_t z_iobuf_wrap(uint8_t *buf, unsigned int capacity)
+z_iobuf_t z_iobuf_wrap(uint8_t *buf, size_t capacity)
 {
     return z_iobuf_wrap_wo(buf, capacity, 0, 0);
 }
 
-z_iobuf_t z_iobuf_make(unsigned int capacity)
+z_iobuf_t z_iobuf_make(size_t capacity)
 {
     return z_iobuf_wrap((uint8_t *)malloc(capacity), capacity);
 }
@@ -46,11 +46,11 @@ void z_iobuf_free(z_iobuf_t *buf)
     buf = 0;
 }
 
-unsigned int z_iobuf_readable(const z_iobuf_t *iob)
+size_t z_iobuf_readable(const z_iobuf_t *iob)
 {
     return iob->w_pos - iob->r_pos;
 }
-unsigned int z_iobuf_writable(const z_iobuf_t *iob)
+size_t z_iobuf_writable(const z_iobuf_t *iob)
 {
     return iob->capacity - iob->w_pos;
 }
@@ -61,14 +61,14 @@ void z_iobuf_write(z_iobuf_t *iob, uint8_t b)
     iob->buf[iob->w_pos++] = b;
 }
 
-void z_iobuf_write_slice(z_iobuf_t *iob, const uint8_t *bs, unsigned int offset, unsigned int length)
+void z_iobuf_write_slice(z_iobuf_t *iob, const uint8_t *bs, size_t offset, size_t length)
 {
     assert(z_iobuf_writable(iob) >= length);
     memcpy(iob->buf + iob->w_pos, bs + offset, length);
     iob->w_pos += length;
 }
 
-void z_iobuf_write_bytes(z_iobuf_t *iob, const uint8_t *bs, unsigned int length)
+void z_iobuf_write_bytes(z_iobuf_t *iob, const uint8_t *bs, size_t length)
 {
     assert(z_iobuf_writable(iob) >= length);
     memcpy(iob->buf + iob->w_pos, bs, length);
@@ -81,7 +81,7 @@ uint8_t z_iobuf_read(z_iobuf_t *iob)
     return iob->buf[iob->r_pos++];
 }
 
-uint8_t *z_iobuf_read_to_n(z_iobuf_t *iob, uint8_t *dst, unsigned int length)
+uint8_t *z_iobuf_read_to_n(z_iobuf_t *iob, uint8_t *dst, size_t length)
 {
     assert(z_iobuf_readable(iob) >= length);
     memcpy(dst, iob->buf + iob->r_pos, length);
@@ -89,19 +89,19 @@ uint8_t *z_iobuf_read_to_n(z_iobuf_t *iob, uint8_t *dst, unsigned int length)
     return dst;
 }
 
-uint8_t *z_iobuf_read_n(z_iobuf_t *iob, unsigned int length)
+uint8_t *z_iobuf_read_n(z_iobuf_t *iob, size_t length)
 {
     uint8_t *dst = (uint8_t *)malloc(length);
     return z_iobuf_read_to_n(iob, dst, length);
 }
 
-void z_iobuf_put(z_iobuf_t *iob, uint8_t b, unsigned int pos)
+void z_iobuf_put(z_iobuf_t *iob, uint8_t b, size_t pos)
 {
     assert(pos < iob->capacity);
     iob->buf[pos] = b;
 }
 
-uint8_t z_iobuf_get(z_iobuf_t *iob, unsigned int pos)
+uint8_t z_iobuf_get(z_iobuf_t *iob, size_t pos)
 {
     assert(pos < iob->capacity);
     return iob->buf[pos];
